tiktok_image_app/backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
moviepy>=1.0.3
gunicorn>=21.2.0

